Move from utils to metrics and rename classification_report.
