+++ power/metrics.py
@@ -4,6 +4,8 @@ from typing import Optional
 import numpy as np
 from scipy import stats
 
+from power.types import PowerOutput
+
 from .compute_power import compute_power
 from .dgps import dgps
 from .effects import effects
@@ -11,7 +13,7 @@ from .stats_tests import stats_tests
 from .types import DGPParameters, PowerOutput
 
 
-def classification_report(
+def estimate_power_from_predictions(
     preds_a: list | np.ndarray,
     preds_b: list | np.ndarray,
     references: list | np.ndarray,
@@ -241,4 +243,109 @@ def find_minimum_detectable_effect(
     return np.sqrt(
         (((z_alpha + z_beta) ** 2) * (omega + (var_a / k_a) + (var_b / k_b)))
         / dataset_size
+    )
+
+
+def make_probability_table(
+    baseline_acc: float = 0.5,  # accuracy of the baseline model (Model 1/A)
+    delta_acc: float = 0.1,  # accuracy difference between Model 1 and Model 2, \theta_B - baseline_acc
+    agreement_rate: (
+        float | None
+    ) = 0.8,  # agreement rate between the two models (Model 1 and Model 2)
+) -> np.ndarray:
+    """
+    Returns a 2x2 probability table representing the joint distribution of two models predictions.
+    The table is structured as follows:
+        [
+            [p_both_incorrect, p_only_2_correct],
+            [p_only_1_correct, p_both_correct]
+        ]
+
+    where:
+      - the baseline model is model 1, so that accuracy_model1 = table[1, :].sum() = baseline_acc
+      - the second model, model 2, has accuracy_model2 = table[:, 1].sum() = baseline_acc + delta_acc
+      - the agreement rate is the sum of the probabilities of both models giving the same prediction, i.e.,
+            table.diagonal.sum() = agreement_rate
+    """
+    # p11 is the probability that both models are correct
+    if agreement_rate is not None:
+        p11 = (agreement_rate + 2 * baseline_acc + delta_acc - 1) / 2
+        p10 = baseline_acc - p11
+        p01 = baseline_acc + delta_acc - p11
+        p00 = agreement_rate - p11
+        probs = np.array([p00, p01, p10, p11])
+    else:
+        acc1 = baseline_acc
+        acc2 = baseline_acc + delta_acc
+        if acc1 + acc2 <= 1.0:
+            # Case A: no overlapping “both correct” mass possible; set p11 = 0
+            p11 = 0.0
+            p10 = acc1  # Pr(model1 correct, model2 wrong)
+            p01 = acc2  # Pr(model1 wrong, model2 correct)
+            p00 = 1.0 - acc1 - acc2
+        else:
+            # Case B: some overlap is forced; set p11 = acc1 + acc2 - 1
+            p11 = acc1 + acc2 - 1.0
+            p10 = acc1 - p11  # = acc1 - (acc1 + acc2 - 1) = 1 - acc2
+            p01 = acc2 - p11  # = acc2 - (acc1 + acc2 - 1) = 1 - acc1
+            p00 = 0.0
+
+        probs = np.array([p00, p01, p10, p11], dtype=float)
+
+    if np.any(probs < 0) or not np.isclose(probs.sum(), 1.0):
+        raise ValueError(
+            "The provided parameters do not yield a valid probability table."
+        )
+
+    table = probs.reshape(2, 2)
+    # some more sanity checks
+    assert np.isclose(table[1, :].sum(), baseline_acc), "Error!"
+    assert np.isclose(table[:, 1].sum(), baseline_acc + delta_acc), "Error!"
+    if agreement_rate is not None:
+        assert np.isclose(table.diagonal().sum(), agreement_rate), "Error!"
+
+    return table
+
+
+def estimate_power_from_accuracy(
+    baseline_acc: float,
+    delta_acc: float,
+    dataset_size: int,
+    agreement: float = 0.75,
+    alpha: float = 0.05,
+    iterations: int = 1000,
+    seed: int = 42,
+    dgp: str = "dgp::contingency_table",
+    test: str = "stats_test::mcnemar",
+    effect: str = "effect::cohens_g",
+) -> PowerOutput:
+    """
+    Estimate statistical power given baseline and delta accuracy.
+
+    Returns:
+        PowerOutput
+    """
+    try:
+        prob_table = make_probability_table(baseline_acc, delta_acc, agreement)
+    except ValueError:
+        return PowerOutput(
+            power=np.nan, mean_eff=np.nan, type_m=np.nan, type_s=np.nan
+        )
+
+    dgp_params = DGPParameters(
+        true_prob_table=prob_table, dataset_size=dataset_size
+    )
+
+    data_generating_fn = partial(dgps.get(dgp), dgp_params=dgp_params)
+    effect_fn = effects.get(effect)
+    statistical_test_fn = partial(stats_tests.get(test), effect_fn=effect_fn)
+    true_effect_fn = partial(effect_fn, dgp_params=dgp_params)
+
+    return compute_power(
+        data_generating_fn=data_generating_fn,
+        hypothesis_test_fn=statistical_test_fn,
+        true_effect_fn=true_effect_fn,
+        iterations=iterations,
+        alpha=alpha,
+        seed=seed,
     )
